[docs] had to rename kern._src to kern.src. PLEASE CHECK YOUR CODE, if it is running smoothly
