+++ GPy/kern/__init__.py
@@ -1,31 +1,31 @@
 """
 Kernel module the kernels to sit in.
 
-.. automodule:: GPy.kern.src
+.. automodule:: .src
    :members:
    :private-members:
 """
-from . import _src as src
-from ._src.kern import Kern
-from ._src.rbf import RBF
-from ._src.linear import Linear, LinearFull
-from ._src.static import Bias, White, Fixed
-from ._src.brownian import Brownian
-from ._src.stationary import Exponential, OU, Matern32, Matern52, ExpQuad, RatQuad, Cosine
-from ._src.mlp import MLP
-from ._src.periodic import PeriodicExponential, PeriodicMatern32, PeriodicMatern52
-from ._src.standard_periodic import StdPeriodic
-from ._src.independent_outputs import IndependentOutputs, Hierarchical
-from ._src.coregionalize import Coregionalize
-from ._src.ODE_UY import ODE_UY
-from ._src.ODE_UYC import ODE_UYC
-from ._src.ODE_st import ODE_st
-from ._src.ODE_t import ODE_t
-from ._src.poly import Poly
-from ._src.eq_ode2 import EQ_ODE2
-from ._src.trunclinear import TruncLinear,TruncLinear_inf
-from ._src.splitKern import SplitKern,DEtime
-from ._src.splitKern import DEtime as DiffGenomeKern
-from ._src.spline import Spline
-from ._src.eq_ode2 import EQ_ODE2
-from ._src.basis_funcs import LinearSlopeBasisFuncKernel, BasisFuncKernel, ChangePointBasisFuncKernel, DomainKernel
+from .src.kern import Kern
+from .src.add import Add
+from .src.prod import Prod
+from .src.rbf import RBF
+from .src.linear import Linear, LinearFull
+from .src.static import Bias, White, Fixed
+from .src.brownian import Brownian
+from .src.stationary import Exponential, OU, Matern32, Matern52, ExpQuad, RatQuad, Cosine
+from .src.mlp import MLP
+from .src.periodic import PeriodicExponential, PeriodicMatern32, PeriodicMatern52
+from .src.standard_periodic import StdPeriodic
+from .src.independent_outputs import IndependentOutputs, Hierarchical
+from .src.coregionalize import Coregionalize
+from .src.ODE_UY import ODE_UY
+from .src.ODE_UYC import ODE_UYC
+from .src.ODE_st import ODE_st
+from .src.ODE_t import ODE_t
+from .src.poly import Poly
+from .src.eq_ode2 import EQ_ODE2
+from .src.trunclinear import TruncLinear,TruncLinear_inf
+from .src.splitKern import SplitKern,DEtime
+from .src.splitKern import DEtime as DiffGenomeKern
+from .src.spline import Spline
+from .src.basis_funcs import LinearSlopeBasisFuncKernel, BasisFuncKernel, ChangePointBasisFuncKernel, DomainKernel
