+++ GPy/kern/src/__init__.py
@@ -1,0 +1,1 @@
+from . import psi_comp
