+++ GPy/kern/src/add.py
