+++ GPy/kern/src/eq_ode2.py
