+++ GPy/kern/src/kernel_slice_operations.py
