+++ GPy/kern/src/poly.py
