+++ GPy/kern/src/prod.py
