+++ GPy/kern/src/psi_comp/linear_psi_comp.py
