+++ GPy/kern/src/psi_comp/sslinear_psi_comp.py
