+++ GPy/kern/src/psi_comp/ssrbf_psi_comp.py
