+++ GPy/kern/src/splitKern.py
@@ -3,7 +3,7 @@ A new kernel
 """
 
 import numpy as np
-from .kern import Kern,CombinationKernel
+from .kern import Kern, CombinationKernel
 from .independent_outputs import index_to_slices
 import itertools
 
@@ -13,7 +13,7 @@ class DEtime(Kern):
         self.idx_p = idx_p
         self.index_dim=index_dim
         self.kern = SplitKern(kernel,Xp, index_dim=index_dim)
-        super(DiffGenomeKern, self).__init__(input_dim=kernel.input_dim+1, active_dims=None, name=name)
+        super(DEtime, self).__init__(input_dim=kernel.input_dim+1, active_dims=None, name=name)
         self.add_parameter(self.kern)
     
     def K(self, X, X2=None):
