+++ GPy/kern/src/standard_periodic.py
