+++ GPy/kern/src/static.py
