+++ GPy/kern/src/stationary.py
