+++ GPy/kern/src/stationary_cython.c
