+++ GPy/kern/src/stationary_cython.pyx
