+++ GPy/kern/src/stationary_utils.c
