+++ GPy/kern/src/stationary_utils.h
