+++ GPy/kern/src/symbolic.py
@@ -1,7 +1,7 @@
 # Check Matthew Rocklin's blog post.
 import sympy as sym
 import numpy as np
-from .kern import Kern
+from GPy.kern.src import Kern
 from ...core.symbolic import Symbolic_core
