+++ GPy/kern/src/sympy_helpers.cpp
