+++ GPy/kern/src/sympy_helpers.h
