+++ GPy/kern/src/todo/eq_ode1.py
