+++ GPy/kern/src/todo/finite_dimensional.py
