+++ GPy/kern/src/todo/gibbs.py
