+++ GPy/kern/src/todo/hetero.py
