+++ GPy/kern/src/todo/poly.py
