+++ GPy/kern/src/todo/rbf_inv.py
