+++ GPy/kern/src/todo/spline.py
