+++ GPy/kern/src/todo/symmetric.py
