+++ setup.py
@@ -76,9 +76,9 @@ else:
     compile_flags = [ '-fopenmp', '-O3', ]
     link_args = ['-lgomp']
 
-ext_mods = [Extension(name='GPy.kern._src.stationary_cython',
-                      sources=['GPy/kern/_src/stationary_cython.c',
-                               'GPy/kern/_src/stationary_utils.c'],
+ext_mods = [Extension(name='GPy.kern.src.stationary_cython',
+                      sources=['GPy/kern/src/stationary_cython.c',
+                               'GPy/kern/src/stationary_utils.c'],
                       include_dirs=[np.get_include(),'.'],
                       extra_compile_args=compile_flags,
                       extra_link_args = link_args),
@@ -91,8 +91,8 @@ ext_mods = [Extension(name='GPy.kern._src.stationary_cython',
                       sources=['GPy/util/linalg_cython.c'],
                       include_dirs=[np.get_include(),'.'],
                       extra_compile_args=compile_flags),
-            Extension(name='GPy.kern._src.coregionalize_cython',
-                      sources=['GPy/kern/_src/coregionalize_cython.c'],
+            Extension(name='GPy.kern.src.coregionalize_cython',
+                      sources=['GPy/kern/src/coregionalize_cython.c'],
                       include_dirs=[np.get_include(),'.'],
                       extra_compile_args=compile_flags)]
 
@@ -109,8 +109,8 @@ setup(name = 'GPy',
                   "GPy.core",
                   "GPy.core.parameterization", 
                   "GPy.kern",
-                  "GPy.kern._src",
-                  "GPy.kern._src.psi_comp", 
+                  "GPy.kern.src",
+                  "GPy.kern.src.psi_comp", 
                   "GPy.models",
                   "GPy.inference",
                   "GPy.inference.optimization",
